Controllerek visszahozatala.
