+++ src/main/java/hu/unideb/inf/ENDSceneController.java
@@ -11,10 +11,10 @@ import java.io.IOException;
 import java.util.Objects;
 
 public class ENDSceneController {
+
     @FXML
     void NextButtonPushed(ActionEvent event) throws IOException {
-
-
+        MainApp.newOrder();
         Parent Next_to_Menu = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("/fxml/Menu.fxml")));
         Scene Menu_Scene = new Scene(Next_to_Menu);
         Stage window = (Stage)((Node)event.getSource()).getScene().getWindow();
