+++ src/main/java/hu/unideb/inf/SummaryController.java
@@ -36,8 +36,8 @@ public class SummaryController implements Initializable{
     @FXML
     void BackToMealPushed(ActionEvent event) throws Exception {
         MainApp.order.setEtel_ital(null);
-        MainApp.order_list.clear();
-        MainApp.order_list_DB.clear();
+        MealSelectController.order_list.clear();
+        MealSelectController.order_list_DB.clear();
 
         Parent Back_to_Meal = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("/fxml/Meal_Select.fxml")));
         Scene Meal_scene = new Scene(Back_to_Meal);
@@ -46,12 +46,14 @@ public class SummaryController implements Initializable{
         window.show();
     }
 
-
     @FXML
     void NextToEndScenePushed(ActionEvent event) throws Exception {
         OrderDAO oDAO = new JpaOrderDAO();
         oDAO.saveOrder(MainApp.order);
         oDAO.close();
+
+        MealSelectController.order_list.clear();
+        MealSelectController.order_list_DB.clear();
 
         Parent Next_to_End = FXMLLoader.load(Objects.requireNonNull(getClass().getResource("/fxml/END_Scene.fxml")));
         Scene End_scene = new Scene(Next_to_End);
@@ -71,39 +73,44 @@ public class SummaryController implements Initializable{
         //Datum és ido
         Time_label.setText(MainApp.order.getDp() + " " + MainApp.order.getIdopont());
 
-        //Etel_ital
-        String[] tomb = MainApp.order.getEtel_ital().split(";");
-        StringBuilder meal = new StringBuilder();
+        if (MealSelectController.order_list_DB.size() != 0) {
+            //Etel_ital
+            String[] tomb = MainApp.order.getEtel_ital().split(";");
+            StringBuilder meal = new StringBuilder();
 
-        Map<String, Integer> map = new TreeMap<>();
+            Map<String, Integer> map = new TreeMap<>();
 
-        for (String s : tomb) {
-            String str = switch (s) {
-                case "c_s" -> "Cola (kicsi)";
-                case "c_m" -> "Cola (közepes)";
-                case "c_l" -> "Cola (nagy)";
-                case "f_s" -> "Fanta (kicsi)";
-                case "f_m" -> "Fanta (közepes)";
-                case "f_l" -> "Fanta (nagy)";
-                case "p_s" -> "Popcorn (kicsi)";
-                case "p_m" -> "Popcorn (közepes)";
-                case "p_l" -> "Popcorn (nagy)";
-                case "n_s" -> "Nachos (kicsi)";
-                case "n_m" -> "Nachos (közepes)";
-                case "n_l" -> "Nachos (nagy)";
-                default -> null;
-            };
-            if (map.containsKey(str)) {
-                int x = map.get(str);
-                x++;
-                map.put(str, x);
-            } else
-                map.put(str, 1);
+            for (String s : tomb) {
+                String str = switch (s) {
+                    case "c_s" -> "Cola (kicsi)";
+                    case "c_m" -> "Cola (közepes)";
+                    case "c_l" -> "Cola (nagy)";
+                    case "f_s" -> "Fanta (kicsi)";
+                    case "f_m" -> "Fanta (közepes)";
+                    case "f_l" -> "Fanta (nagy)";
+                    case "p_s" -> "Popcorn (kicsi)";
+                    case "p_m" -> "Popcorn (közepes)";
+                    case "p_l" -> "Popcorn (nagy)";
+                    case "n_s" -> "Nachos (kicsi)";
+                    case "n_m" -> "Nachos (közepes)";
+                    case "n_l" -> "Nachos (nagy)";
+                    default -> null;
+                };
+                if (map.containsKey(str)) {
+                    int x = map.get(str);
+                    x++;
+                    map.put(str, x);
+                } else
+                    map.put(str, 1);
+            }
+
+            for (Map.Entry<String, Integer> entry : map.entrySet()) {
+                meal.append(entry.getValue()).append(" * ").append(entry.getKey()).append("\n");
+            }
+            Meal_label.setText(meal.toString());
         }
-        for (Map.Entry<String, Integer> entry : map.entrySet()) {
-            meal.append(entry.getValue()).append(" * ").append(entry.getKey()).append("\n");
-        }
-        Meal_label.setText(meal.toString());
+        else
+            Meal_label.setText("Nem rendelt.");
 
         //Ülőhely
         String[] tomb2 = MainApp.order.getUlo_hely().split(";");
